+++ tumeplayMobile/src/components/LevelPointsIndicator.js
@@ -1,21 +1,30 @@
+import {PossibleFragmentSpreadsRule} from 'graphql';
 import React, {useContext} from 'react';
 import {View, Text, StyleSheet, Image, TouchableOpacity} from 'react-native';
 import AppContext from '../../AppContext';
-import coin from '../assets/coin.png';
 import {Colors} from '../styles/Style';
+import * as Progress from 'react-native-progress';
 
 const LevelPointsIndicator = ({style, onPress}) => {
   const context = useContext(AppContext);
+  const {user, doneModules_ids} = context;
 
-  const points = context.points;
+  const fullProgressLength = 6;
+
+  const progress = doneModules_ids.length / fullProgressLength;
 
   return (
     <TouchableOpacity style={[style, styles.container]} onPress={onPress}>
-      <Text style={styles.level}>NIVEAU 1</Text>
-      <View style={styles.pointsContainer}>
-        <Text style={styles.text}>{points} / 3000 </Text>
-        <Image source={coin} style={styles.image} />
-      </View>
+      <Text style={styles.text}>
+        Niveau <Text style={styles.boldText}>{user.level}</Text>
+      </Text>
+      <Progress.Bar
+        progress={progress}
+        width={150}
+        color={'#51B070'}
+        unfilledColor="#DFD7CD"
+        borderColor="#DFD7CD"
+      />
     </TouchableOpacity>
   );
 };
@@ -24,34 +33,29 @@ const styles = StyleSheet.create({
   container: {
     height: 50,
     display: 'flex',
-    flexDirection: 'row',
+    flexDirection: 'column',
     justifyContent: 'space-between',
     alignItems: 'center',
-    borderColor: '#ECC160',
-    borderWidth: 2,
-    borderRadius: 4,
-    backgroundColor: '#FFF',
     paddingHorizontal: 15,
     paddingVertical: 10,
   },
-  pointsContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
   text: {
     fontWeight: 'bold',
-
     paddingRight: 5,
+    fontSize: 20,
+    lineHeight: 22,
+    marginBottom: 6,
   },
   level: {
     paddingRight: 15,
     fontWeight: 'bold',
   },
-  image: {
-    width: 15,
-    height: 15,
+  boldText: {
+    fontWeight: '700',
+    paddingRight: 5,
+    fontSize: 20,
+    lineHeight: 22,
+    marginBottom: 6,
   },
 });