+++ tumeplayMobile/src/components/Quizz/QuizzAnswerButton.js
@@ -4,32 +4,31 @@ import config from '../../../config';
 import Text from '../../components/Text';
 
 const QuizzAnswerButton = props => {
-  const {answer, onPress, name, correctAnswer, hasAnswered, disabled} = props;
+  const {
+    answer,
+    onPress,
+    name,
+    correctAnswer,
+    hasAnswered,
+    disabled,
+    answeredKey,
+  } = props;
+
   return (
     <TouchableOpacity disabled={disabled} onPress={onPress} name={name}>
       <View
         style={[
           styles.button,
-          hasAnswered && answer.key === correctAnswer
-            ? styles.correctAnswer
-            : '',
+          hasAnswered && answeredKey === answer.key && styles.wrongAnswer,
+          hasAnswered && answer.key === correctAnswer && styles.correctAnswer,
         ]}>
         <Text
           style={[
             styles.value,
-            hasAnswered && answer.key === correctAnswer
-              ? styles.correctAnswer
-              : '',
+            hasAnswered && answer.key === correctAnswer && {fontWeight: '700'},
           ]}>
           {answer.value}
         </Text>
-        {hasAnswered ? (
-          answer.key === correctAnswer ? (
-            <Text style={styles.emoji}>✅</Text>
-          ) : (
-            <Text style={styles.emoji}>❌</Text>
-          )
-        ) : null}
       </View>
     </TouchableOpacity>
   );
@@ -50,8 +49,13 @@ const styles = StyleSheet.create({
     backgroundColor: '#F3E1E8',
   },
   correctAnswer: {
-    backgroundColor: '#fff',
+    borderColor: '#51B070',
+    borderWidth: 1,
+    backgroundColor: '#DDF4ED',
     fontWeight: '600',
+  },
+  wrongAnswer: {
+    backgroundColor: '#FFF',
   },
   value: {
     justifyContent: 'center',