+++ tumeplayMobile/src/components/Quizz/QuizzModule.js
@@ -24,6 +24,7 @@ const QuizzModule = ({navigation, route}) => {
   const retry = route?.params?.retry;
   const [disabled, setDisabled] = useState(false);
   const [showAnswer, setshowAnswer] = useState(false);
+  const [answeredKey, setAnswerKey] = useState('');
 
   const formatAnswers = () => {
     let tmpResponses = [];
@@ -35,14 +36,15 @@ const QuizzModule = ({navigation, route}) => {
     setResponses(tmpResponses);
   };
 
-  const displayAnswerText = ans => {
-    if (ans === responses[responses?.length - 1]?.value) {
+  const displayAnswerText = answerKey => {
+    if (answerKey === responses[responses?.length - 1]?.value) {
       correctAnswers.push(question);
       setCorrectAnswers([...correctAnswers]);
     } else {
       wrongAnswers.push(question);
       setWrongAnswers([...wrongAnswers]);
     }
+    setAnswerKey(answerKey);
     setDisabled(!disabled);
     setHasAnswered(!hasAnswered);
     setDisplayResponse(!displayResponse);
@@ -57,6 +59,7 @@ const QuizzModule = ({navigation, route}) => {
           hasAnswered={hasAnswered}
           disabled={disabled}
           key={ans.key}
+          answeredKey={answeredKey}
           onPress={() => displayAnswerText(ans.key)}
         />
       );